Pipe: refactor package structure (db-test)
