+++ server/src/test/java/org/apache/iotdb/db/pipe/execution/PipeProcessorSubtaskExecutorTest.java
@@ -17,28 +17,31 @@
  * under the License.
  */
 
-package org.apache.iotdb.db.pipe.execution.executor;
+package org.apache.iotdb.db.pipe.execution;
 
-import org.apache.iotdb.db.pipe.task.connection.BoundedBlockingPendingQueue;
-import org.apache.iotdb.db.pipe.task.subtask.PipeConnectorSubtask;
-import org.apache.iotdb.pipe.api.PipeConnector;
+import org.apache.iotdb.db.pipe.execution.executor.PipeProcessorSubtaskExecutor;
+import org.apache.iotdb.db.pipe.task.connection.EventSupplier;
+import org.apache.iotdb.db.pipe.task.subtask.PipeProcessorSubtask;
+import org.apache.iotdb.pipe.api.PipeProcessor;
+import org.apache.iotdb.pipe.api.collector.EventCollector;
 
 import org.junit.Before;
 import org.mockito.Mockito;
 
 import static org.mockito.Mockito.mock;
 
-public class PipeConnectorSubtaskExecutorTest extends PipeSubtaskExecutorTest {
+public class PipeProcessorSubtaskExecutorTest extends PipeSubtaskExecutorTest {
 
   @Before
   public void setUp() throws Exception {
-    executor = new PipeConnectorSubtaskExecutor();
+    executor = new PipeProcessorSubtaskExecutor();
 
     subtask =
         Mockito.spy(
-            new PipeConnectorSubtask(
-                "PipeConnectorSubtaskExecutorTest",
-                mock(BoundedBlockingPendingQueue.class),
-                mock(PipeConnector.class)));
+            new PipeProcessorSubtask(
+                "PipeProcessorSubtaskExecutorTest",
+                mock(EventSupplier.class),
+                mock(PipeProcessor.class),
+                mock(EventCollector.class)));
   }
 }
