+++ server/src/test/java/org/apache/iotdb/db/pipe/execution/PipeSubtaskExecutorTest.java
@@ -17,8 +17,9 @@
  * under the License.
  */
 
-package org.apache.iotdb.db.pipe.execution.executor;
+package org.apache.iotdb.db.pipe.execution;
 
+import org.apache.iotdb.db.pipe.execution.executor.PipeSubtaskExecutor;
 import org.apache.iotdb.db.pipe.task.subtask.PipeSubtask;
 
 import org.junit.After;
